Refactor event-converter and event-commons files- Capture: event-set sub-component #44
