+++ src/app/features/components/capture/current-event/event-type-selecting/event-type-selecting.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {MatRadioChange, MatRadioModule} from '@angular/material/radio';
 import { Logger } from '../../../../../shared/services/logging/logger';
-import { EventTypeSelectingUIModel } from '../../model/capture/ui-model/current-event-ui-model';
+import { EventTypeSelectingUIModel } from "../../model/capture/ui-model/event-type-selecting-ui-model";
 
 
 @Component({
