+++ src/app/features/components/capture/model/capture/business-logic-model/running-events-business-logic-model.ts
@@ -4,7 +4,7 @@ import { Observable, Subject } from "rxjs";
 import { IEvent } from "../capture-common-interfaces";
 import { Logger } from "../../../../../../shared/services/logging/logger";
 import { IPersistedRunningEvents, RunningEventsPersistence } from "../../../../../../shared/classes/db/running-events-db";
-import { encodePersistedEvent } from "./event-converter";
+import { encodePersistedEvent } from "./event-commons";
 
 export interface IRunningEventsBusinessLogicModel {
     readRunninfEventsFromDB(): unknown;
